webapp/requirements_web.txt
Flask
Pillow
numpy
tensorflow
opencv-python
matplotlib
scikit-learn
Werkzeug
Jinja2
gunicorn
python-dotenv
